requirements.txt
# Core Framework
flask==3.0.0
streamlit==1.31.0
gunicorn==21.2.0

# LangChain & LLMs
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-google-genai==0.0.11
langchain-groq==0.0.1
langchain-huggingface==0.0.1
langchain-ollama==0.0.1
langchain-openai==0.0.8

# Multi-Agent Frameworks
crewai==0.11.0
pyautogen==0.2.0
langgraph==0.0.26

# Vector Database & RAG
chromadb==0.4.22
sentence-transformers==2.5.1

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Kaggle Integration
kaggle==1.6.6

# API Clients
groq==0.4.2
google-generativeai==0.3.2
openai==1.12.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
requests==2.31.0
aiohttp==3.9.3

# CORS
flask-cors==4.0.0

# Production Server
waitress==3.0.0





